+++ roslibrust_codegen/src/gen.rs
@@ -1,11 +1,12 @@
 use proc_macro2::TokenStream;
-use quote::{format_ident, quote};
-use std::collections::HashMap;
+use quote::{format_ident, quote, ToTokens};
+use serde::de::DeserializeOwned;
 use std::str::FromStr;
 use syn::parse_quote;
 
 use crate::parse::{
-    ParsedMessageFile, ParsedServiceFile, ROS_2_TYPE_TO_RUST_TYPE_MAP, ROS_TYPE_TO_RUST_TYPE_MAP,
+    convert_ros_type_to_rust_type, ConstantInfo, FieldInfo, ParsedMessageFile, ParsedServiceFile,
+    RosLiteral,
 };
 use crate::utils::RosVersion;
 
@@ -45,47 +46,12 @@ pub fn generate_service(service: ParsedServiceFile) -> TokenStream {
 }
 
 pub fn generate_struct(msg: ParsedMessageFile) -> TokenStream {
-    let msg = replace_ros_types_with_rust_types(msg);
     let attrs = derive_attrs();
     let fields = msg
         .fields
         .into_iter()
-        .map(|mut field| {
-            field.field_type.field_type = match field.field_type.package_name {
-                Some(ref pkg) => {
-                    if pkg.as_str() == msg.package.as_str() {
-                        format!("self::{}", field.field_type.field_type)
-                    } else {
-                        format!("{}::{}", pkg, field.field_type.field_type)
-                    }
-                }
-                None => field.field_type.field_type.clone(),
-            };
-            let field_type = if field.field_type.is_vec {
-                format!("::std::vec::Vec<{}>", field.field_type.field_type)
-            } else {
-                field.field_type.field_type.clone()
-            };
-            let field_type = TokenStream::from_str(field_type.as_str()).unwrap();
-
-            let field_name = format_ident!("r#{}", field.field_name);
-            if let Some(ref default_val) = field.default {
-                if field.field_type.is_vec {
-                    // For vectors use smart_defaults "dynamic" style
-                    quote! {
-                        #[default(_code = #default_val)]
-                        pub #field_name: #field_type,
-                    }
-                } else {
-                    // For non vectors use smart_default's constant style
-                    quote! {
-                      #[default(#default_val)]
-                      pub #field_name: #field_type,
-                    }
-                }
-            } else {
-                quote! { pub #field_name: #field_type, }
-            }
+        .map(|field| {
+            generate_field_definition(field, &msg.package, msg.version.unwrap_or(RosVersion::ROS1))
         })
         .collect::<Vec<TokenStream>>();
 
@@ -93,15 +59,7 @@ pub fn generate_struct(msg: ParsedMessageFile) -> TokenStream {
         .constants
         .into_iter()
         .map(|constant| {
-            let constant_name = format_ident!("r#{}", constant.constant_name);
-            let constant_type = if constant.constant_type == "::std::string::String" {
-                String::from("&'static str")
-            } else {
-                constant.constant_type
-            };
-            let constant_type = TokenStream::from_str(constant_type.as_str()).unwrap();
-            let constant_value = constant.constant_value;
-            quote! { pub const #constant_name: #constant_type = #constant_value; }
+            generate_constant_field_definition(constant, msg.version.unwrap_or(RosVersion::ROS1))
         })
         .collect::<Vec<TokenStream>>();
 
@@ -131,6 +89,68 @@ pub fn generate_struct(msg: ParsedMessageFile) -> TokenStream {
     base
 }
 
+fn generate_field_definition(field: FieldInfo, msg_pkg: &str, version: RosVersion) -> TokenStream {
+    let rust_field_type = match field.field_type.package_name {
+        Some(ref pkg) => {
+            if pkg.as_str() == msg_pkg {
+                format!("self::{}", field.field_type.field_type)
+            } else {
+                format!("{}::{}", pkg, field.field_type.field_type)
+            }
+        }
+        None => convert_ros_type_to_rust_type(version, &field.field_type.field_type)
+            .unwrap_or_else(|| panic!("No Rust type for {}", field.field_type))
+            .to_owned(),
+    };
+    let rust_field_type = if field.field_type.is_vec {
+        format!("::std::vec::Vec<{}>", rust_field_type)
+    } else {
+        rust_field_type
+    };
+    let rust_field_type = TokenStream::from_str(rust_field_type.as_str()).unwrap();
+
+    let field_name = format_ident!("r#{}", field.field_name);
+    if let Some(ref default_val) = field.default {
+        let default_val = ros_literal_to_rust_literal(
+            &field.field_type.field_type,
+            default_val,
+            field.field_type.is_vec,
+        );
+        if field.field_type.is_vec {
+            // For vectors use smart_defaults "dynamic" style
+            quote! {
+                #[default(_code = #default_val)]
+                pub #field_name: #rust_field_type,
+            }
+        } else {
+            // For non vectors use smart_default's constant style
+            quote! {
+              #[default(#default_val)]
+              pub #field_name: #rust_field_type,
+            }
+        }
+    } else {
+        quote! { pub #field_name: #rust_field_type, }
+    }
+}
+
+fn generate_constant_field_definition(constant: ConstantInfo, version: RosVersion) -> TokenStream {
+    let constant_name = format_ident!("r#{}", constant.constant_name);
+    let constant_rust_type =
+        convert_ros_type_to_rust_type(version, &constant.constant_type).unwrap();
+    let constant_rust_type = if constant_rust_type == "::std::string::String" {
+        String::from("&'static str")
+    } else {
+        // Oof it's ugly in here
+        constant_rust_type.to_owned()
+    };
+    let constant_rust_type = TokenStream::from_str(constant_rust_type.as_str()).unwrap();
+    let constant_value =
+        ros_literal_to_rust_literal(&constant.constant_type, &constant.constant_value, false);
+
+    quote! { pub const #constant_name: #constant_rust_type = #constant_value; }
+}
+
 pub fn generate_mod(
     pkg_name: String,
     struct_definitions: Vec<TokenStream>,
@@ -153,50 +173,74 @@ pub fn generate_mod(
     }
 }
 
-pub fn replace_ros_types_with_rust_types(mut msg: ParsedMessageFile) -> ParsedMessageFile {
-    const INTERNAL_STD_MSGS: [&str; 1] = ["Header"];
+fn ros_literal_to_rust_literal(ros_type: &str, literal: &RosLiteral, is_vec: bool) -> TokenStream {
+    // TODO: The naming of all the functions under this tree seems inaccurate
+    parse_ros_value(ros_type, &literal.inner, is_vec)
+}
 
-    // Select which type conversion map to use depending on ros version
-    let prop_map: &HashMap<&'static str, &'static str> = match msg.version {
-        Some(RosVersion::ROS1) => &ROS_TYPE_TO_RUST_TYPE_MAP,
-        Some(RosVersion::ROS2) => &ROS_2_TYPE_TO_RUST_TYPE_MAP,
-        None => {
-            // If we couldn't determine the package type, assume ROS1 for now
-            &ROS_TYPE_TO_RUST_TYPE_MAP
+// Converts a ROS string to a literal value
+// Not intended to be called directly, but only via parse_ros_value.
+// Wraps a serde_json deserialize call with our style of error handling.
+fn generic_parse_value<T: DeserializeOwned + ToTokens + std::fmt::Debug>(
+    value: &str,
+    is_vec: bool,
+) -> TokenStream {
+    if is_vec {
+        let parsed: Vec<T> = serde_json::from_str(value).unwrap_or_else(|_|
+            panic!("Failed to parse a literal value in a message file to the corresponding rust type: {value} to {}", std::any::type_name::<T>())
+        );
+        let vec_str = format!("vec!{parsed:?}");
+        quote! { #vec_str }
+    } else {
+        let parsed: T = serde_json::from_str(value).unwrap_or_else(|_|
+            panic!("Failed to parse a literal value in a message file to the corresponding rust type: {value} to {}", std::any::type_name::<T>())
+        );
+        quote! { #parsed }
+    }
+}
+
+/// For a given, which is either a ROS constant or default, parse the constant and convert it into a rust TokenStream
+/// which represents the same literal value. This handles frustrating edge cases that are not well documented features
+/// in either ROS1 or ROS2 such as:
+/// - `f32[] MY_CONST_ARRAY=[0, 1]` we have to convert the value of the constant to `vec![0.0, 1.0]`
+/// - `string[] names_field ['first', "second"]` we have to convert the default value to `vec!["first".to_string(), "second".to_string()]
+/// Note: I could find NO documentation from ROS about what was actually legal or expected for these value expressions
+/// Note: ROS says "strings are not escaped". No idea how the eff I'm actually supposed to handle that so likely bugs...
+/// Note: No idea of "constant arrays" are intended to be supported in ROS...
+/// `ros_type` -- Expects the string key of the determined rust type to hold the value. Should come from one of the type map constants.
+/// `value` -- Expects the trimmed string containing only the value expression
+/// `is_vec` -- True iff the type is an array type
+/// TODO I'd like this to take FieldType, but want it to also work with constants...
+fn parse_ros_value(ros_type: &str, value: &str, is_vec: bool) -> TokenStream {
+    match ros_type {
+        "bool" => generic_parse_value::<bool>(value, is_vec),
+        "float64" => generic_parse_value::<f64>(value, is_vec),
+        "float32" => generic_parse_value::<f32>(value, is_vec),
+        "uint8" | "char" | "byte" => generic_parse_value::<u8>(value, is_vec),
+        "int8" => generic_parse_value::<i8>(value, is_vec),
+        "uint16" => generic_parse_value::<u16>(value, is_vec),
+        "int16" => generic_parse_value::<i16>(value, is_vec),
+        "uint32" => generic_parse_value::<u32>(value, is_vec),
+        "int32" => generic_parse_value::<i32>(value, is_vec),
+        "uint64" => generic_parse_value::<u64>(value, is_vec),
+        "int64" => generic_parse_value::<i64>(value, is_vec),
+        "string" => {
+            // String is a special case because of quotes and to_string()
+            if is_vec {
+                // TODO there is a bug here, no idea how I should be attempting to convert / escape single quotes here...
+                let parsed: Vec<String> = serde_json::from_str(value).unwrap_or_else(|_|
+            panic!("Failed to parse a literal value in a message file to the corresponding rust type: {value} to Vec<String>"));
+                let vec_str = format!("{parsed:?}.iter().map(|x| x.to_string()).collect()");
+                quote! { #vec_str }
+            } else {
+                // Halfass attempt to deal with ROS's string escaping / quote bullshit
+                let value = &value.replace('\'', "\"");
+                let parsed: String = serde_json::from_str(value).unwrap_or_else(|_| panic!("Failed to parse a literal value in a message file to the corresponding rust type: {value} to String"));
+                quote! { #parsed }
+            }
         }
-    };
-
-    msg.constants = msg
-        .constants
-        .into_iter()
-        .map(|mut constant| {
-            if prop_map.contains_key(constant.constant_type.as_str()) {
-                constant.constant_type = prop_map
-                    .get(constant.constant_type.as_str())
-                    .unwrap()
-                    .to_string();
-                // We do not need to consider the package for constants as they're required
-                // to be built-in types other than Time and Duration (I think Header is not
-                // technically built-in)
-            }
-            constant
-        })
-        .collect();
-    msg.fields = msg
-        .fields
-        .into_iter()
-        .map(|mut field| {
-            field.field_type.field_type = prop_map
-                .get(field.field_type.field_type.as_str())
-                .unwrap_or(&field.field_type.field_type.as_str())
-                .to_string();
-            for std_msg in INTERNAL_STD_MSGS {
-                if field.field_type.field_type.as_str() == std_msg {
-                    field.field_type.package_name = Some("std_msgs".into());
-                }
-            }
-            field
-        })
-        .collect();
-    msg
+        _ => {
+            panic!("Found default for type which does not support default: {ros_type}");
+        }
+    }
 }