+++ roslibrust_codegen/src/parse.rs
@@ -1,8 +1,6 @@
-use proc_macro2::TokenStream;
-use quote::{quote, ToTokens};
-use serde::de::DeserializeOwned;
 use std::{
     collections::HashMap,
+    fmt::Display,
     path::{Path, PathBuf},
 };
 
@@ -26,7 +24,6 @@ lazy_static::lazy_static! {
         ("string", "::std::string::String"),
         ("time", "::roslibrust_codegen::integral_types::Time"),
         ("duration", "::roslibrust_codegen::integral_types::Duration"),
-        ("Header", "Header"),
     ].into_iter().collect();
 
     pub static ref ROS_2_TYPE_TO_RUST_TYPE_MAP: HashMap<&'static str, &'static str> = vec![
@@ -50,6 +47,38 @@ lazy_static::lazy_static! {
     ].into_iter().collect();
 }
 
+pub fn is_intrinsic_type(version: RosVersion, ros_type: &str) -> bool {
+    match version {
+        RosVersion::ROS1 => ROS_TYPE_TO_RUST_TYPE_MAP.contains_key(ros_type),
+        RosVersion::ROS2 => ROS_2_TYPE_TO_RUST_TYPE_MAP.contains_key(ros_type),
+    }
+}
+
+pub fn convert_ros_type_to_rust_type(version: RosVersion, ros_type: &str) -> Option<&'static str> {
+    match version {
+        RosVersion::ROS1 => ROS_TYPE_TO_RUST_TYPE_MAP.get(ros_type).copied(),
+        RosVersion::ROS2 => ROS_2_TYPE_TO_RUST_TYPE_MAP.get(ros_type).copied(),
+    }
+}
+
+/// Stores the ROS string representation of a literal
+#[derive(Clone, Debug)]
+pub struct RosLiteral {
+    pub inner: String,
+}
+
+impl std::fmt::Display for RosLiteral {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        self.inner.fmt(f)
+    }
+}
+
+impl From<String> for RosLiteral {
+    fn from(value: String) -> Self {
+        Self { inner: value }
+    }
+}
+
 /// Describes the type for an individual field in a message
 #[derive(PartialEq, Eq, Hash, Debug, Clone)]
 pub struct FieldType {
@@ -63,14 +92,23 @@ pub struct FieldType {
     pub is_vec: bool,
 }
 
+impl Display for FieldType {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        if self.is_vec {
+            f.write_fmt(format_args!("{}[]", self.field_type))
+        } else {
+            f.write_fmt(format_args!("{}", self.field_type))
+        }
+    }
+}
+
 /// Describes all information for an individual field
 #[derive(Clone, Debug)]
 pub struct FieldInfo {
     pub field_type: FieldType,
     pub field_name: String,
     // Exists if this is a ros2 message field with a default value
-    // Contains a literal token representing the value, and bool indicating if the value is static.
-    pub default: Option<TokenStream>,
+    pub default: Option<RosLiteral>,
 }
 
 // Because TokenStream doesn't impl PartialEq we have to do it manually for FieldInfo
@@ -87,7 +125,7 @@ impl PartialEq for FieldInfo {
 pub struct ConstantInfo {
     pub constant_type: String,
     pub constant_name: String,
-    pub constant_value: TokenStream,
+    pub constant_value: RosLiteral,
 }
 
 // Because TokenStream doesn't impl PartialEq we have to do it manually for ConstantInfo
@@ -120,6 +158,10 @@ impl ParsedMessageFile {
                     || field.field_type.package_name == Some(String::from("std_msgs")))
         })
     }
+
+    pub fn get_full_name(&self) -> String {
+        format!("{}/{}", self.package, self.name)
+    }
 }
 
 /// Describes all information for a single service file
@@ -141,13 +183,13 @@ pub struct ParsedServiceFile {
 fn parse_field(line: &str, pkg: &Package, msg_name: &str) -> FieldInfo {
     let mut splitter = line.split_whitespace();
     let pkg_name = pkg.name.as_str();
-    let field_type = splitter.next().expect(&format!(
-        "Did not find field_type on line: {line} while parsing {pkg_name}/{msg_name}"
-    ));
+    let field_type = splitter.next().unwrap_or_else(|| {
+        panic!("Did not find field_type on line: {line} while parsing {pkg_name}/{msg_name}")
+    });
     let field_type = parse_type(field_type, pkg);
-    let field_name = splitter.next().expect(&format!(
-        "Did not find field_name on line: {line} while parsing {pkg_name}/{msg_name}"
-    ));
+    let field_name = splitter.next().unwrap_or_else(|| {
+        panic!("Did not find field_name on line: {line} while parsing {pkg_name}/{msg_name}")
+    });
 
     let sep = line.find(' ').unwrap();
     // Determine if there is a default value for this field
@@ -160,11 +202,7 @@ fn parse_field(line: &str, pkg: &Package, msg_name: &str) -> FieldInfo {
                 if remainder.is_empty() {
                     None
                 } else {
-                    Some(parse_ros_value(
-                        &field_type.field_type,
-                        remainder,
-                        field_type.is_vec,
-                    ))
+                    Some(remainder.to_owned().into())
                 }
             }
             None => {
@@ -185,7 +223,7 @@ fn parse_field(line: &str, pkg: &Package, msg_name: &str) -> FieldInfo {
 
 fn parse_constant_field(line: &str, pkg: &Package) -> ConstantInfo {
     let sep = line.find(' ').unwrap();
-    let equal_after_sep = line[sep..].find("=").unwrap();
+    let equal_after_sep = line[sep..].find('=').unwrap();
     let mut constant_type = parse_type(line[..sep].trim(), pkg).field_type;
     let constant_name = line[sep + 1..(equal_after_sep + sep)].trim().to_string();
 
@@ -195,12 +233,10 @@ fn parse_constant_field(line: &str, pkg: &Package) -> ConstantInfo {
     }
 
     let constant_value = line[sep + equal_after_sep + 1..].trim().to_string();
-    // TODO bug here, explicitly not handling constant vec until we can manage static array generation
-    let constant_value = parse_ros_value(&constant_type, &constant_value, false);
     ConstantInfo {
         constant_type,
         constant_name,
-        constant_value,
+        constant_value: constant_value.into(),
     }
 }
 
@@ -220,16 +256,16 @@ pub fn parse_ros_message_file(
 
     for line in data.lines() {
         let line = strip_comments(line).trim();
-        if line.len() == 0 {
+        if line.is_empty() {
             // Comment only line skip
             continue;
         }
         // Determine if we're looking at a constant or a field
-        let sep = line.find(' ').expect(&format!(
-            "Found an invalid ros field line, no space delinting type from name: {line}"
-        ));
-        let equal_after_sep = line[sep..].find("=");
-        if let Some(_) = equal_after_sep {
+        let sep = line.find(' ').unwrap_or_else(|| {
+            panic!("Found an invalid ros field line, no space delinting type from name: {line}")
+        });
+        let equal_after_sep = line[sep..].find('=');
+        if equal_after_sep.is_some() {
             // Since we found an equal sign after a space, this must be a constant
             constants.push(parse_constant_field(line, package))
         } else {
@@ -238,8 +274,8 @@ pub fn parse_ros_message_file(
         }
     }
     ParsedMessageFile {
-        fields: fields,
-        constants: constants,
+        fields,
+        constants,
         name: name.to_owned(),
         package: package.name.clone(),
         version: package.version,
@@ -261,7 +297,7 @@ pub fn parse_ros_service_file(
 ) -> ParsedServiceFile {
     let mut dash_line_number = None;
     for (line_num, line) in data.lines().enumerate() {
-        match (line.find("---"), line.find("#")) {
+        match (line.find("---"), line.find('#')) {
             (Some(dash_idx), Some(cmt_idx)) => {
                 if dash_idx < cmt_idx {
                     // Comment appears after dash
@@ -278,17 +314,16 @@ pub fn parse_ros_service_file(
     }
     let str_accumulator = |mut acc: String, line: &str| -> String {
         acc.push_str(line);
-        acc.push_str("\n");
+        acc.push('\n');
         acc
     };
 
-    let dash_line_number = dash_line_number.expect(
-        format!(
+    let dash_line_number = dash_line_number.unwrap_or_else(|| {
+        panic!(
             "Failed to find delimiter line '---' in {}/{name}",
             &package.name
         )
-        .as_str(),
-    );
+    });
     let request_str = data
         .lines()
         .take(dash_line_number)
@@ -331,29 +366,20 @@ fn strip_comments(line: &str) -> &str {
 fn parse_field_type(type_str: &str, is_vec: bool, pkg: &Package) -> FieldType {
     let items = type_str.split('/').collect::<Vec<&str>>();
 
-    // Select which type conversion map to use depending on package version
-    let prop_map: &HashMap<&'static str, &'static str> = match pkg.version {
-        Some(RosVersion::ROS1) => &ROS_TYPE_TO_RUST_TYPE_MAP,
-        Some(RosVersion::ROS2) => &ROS_2_TYPE_TO_RUST_TYPE_MAP,
-        None => {
-            // If we couldn't determine the package type, assume ROS1 for now
-            &ROS_TYPE_TO_RUST_TYPE_MAP
-        }
-    };
-
     if items.len() == 1 {
         // If there is only one item (no package redirect)
+        let pkg_version = pkg.version.unwrap_or(RosVersion::ROS1);
         FieldType {
-            package_name: if prop_map.contains_key(type_str) {
+            package_name: if is_intrinsic_type(pkg_version, type_str) {
                 // If it is a fundamental type, no package
-                if items[0] == "Header" {
-                    Some("std_msgs".to_owned())
-                } else {
-                    None
-                }
+                None
             } else {
                 // Otherwise it is referencing another message in the same package
-                Some(pkg.name.clone())
+                if type_str == "Header" {
+                    Some("std_msgs".to_owned())
+                } else {
+                    Some(pkg.name.clone())
+                }
             },
             field_type: items[0].to_string(),
             is_vec,
@@ -383,8 +409,8 @@ fn parse_field_type(type_str: &str, is_vec: bool, pkg: &Package) -> FieldType {
 /// `pkg` -- Reference to package this type is within, used for version information and determining relative types
 fn parse_type(type_str: &str, pkg: &Package) -> FieldType {
     // Handle array logic
-    let open_bracket_idx = type_str.find("[");
-    let close_bracket_idx = type_str.find("]");
+    let open_bracket_idx = type_str.find('[');
+    let close_bracket_idx = type_str.find(']');
     match (open_bracket_idx, close_bracket_idx) {
         (Some(o), Some(_c)) => {
             // After having stripped array information, parse the remainder of the type
@@ -396,73 +422,6 @@ fn parse_type(type_str: &str, pkg: &Package) -> FieldType {
         }
         _ => {
             panic!("Found malformed type: {type_str}");
-        }
-    }
-}
-
-// Converts a ROS string to a literal value
-// Not intended to be called directly, but only via parse_ros_value.
-// Wraps a serde_json deserialize call with our style of error handling.
-fn generic_parse_value<T: DeserializeOwned + ToTokens + std::fmt::Debug>(
-    value: &str,
-    is_vec: bool,
-) -> TokenStream {
-    if is_vec {
-        let parsed: Vec<T> = serde_json::from_str(value).expect(
-            &format!("Failed to parse a literal value in a message file to the corresponding rust type: {value} to {}", std::any::type_name::<T>())
-        );
-        let vec_str = format!("vec!{parsed:?}");
-        quote! { #vec_str }
-    } else {
-        let parsed: T = serde_json::from_str(value).expect(
-            &format!("Failed to parse a literal value in a message file to the corresponding rust type: {value} to {}", std::any::type_name::<T>())
-        );
-        quote! { #parsed }
-    }
-}
-
-/// For a given, which is either a ROS constant or default, parse the constant and convert it into a rust TokenStream
-/// which represents the same literal value. This handles frustrating edge cases that are not well documented features
-/// in either ROS1 or ROS2 such as:
-/// - `f32[] MY_CONST_ARRAY=[0, 1]` we have to convert the value of the constant to `vec![0.0, 1.0]`
-/// - `string[] names_field ['first', "second"]` we have to convert the default value to `vec!["first".to_string(), "second".to_string()]
-/// Note: I could find NO documentation from ROS about what was actually legal or expected for these value expressions
-/// Note: ROS says "strings are not escaped". No idea how the eff I'm actually supposed to handle that so likely bugs...
-/// Note: No idea of "constant arrays" are intended to be supported in ROS...
-/// `ros_type` -- Expects the string key of the determined rust type to hold the value. Should come from one of the type map constants.
-/// `value` -- Expects the trimmed string containing only the value expression
-/// `is_vec` -- True iff the type is an array type
-/// TODO I'd like this to take FieldType, but want it to also work with constants...
-fn parse_ros_value(ros_type: &str, value: &str, is_vec: bool) -> TokenStream {
-    match ros_type {
-        "bool" => generic_parse_value::<bool>(value, is_vec),
-        "float64" => generic_parse_value::<f64>(value, is_vec),
-        "float32" => generic_parse_value::<f32>(value, is_vec),
-        "uint8" | "char" | "byte" => generic_parse_value::<u8>(value, is_vec),
-        "int8" => generic_parse_value::<i8>(value, is_vec),
-        "uint16" => generic_parse_value::<u16>(value, is_vec),
-        "int16" => generic_parse_value::<i16>(value, is_vec),
-        "uint32" => generic_parse_value::<u32>(value, is_vec),
-        "int32" => generic_parse_value::<i32>(value, is_vec),
-        "uint64" => generic_parse_value::<u64>(value, is_vec),
-        "int64" => generic_parse_value::<i64>(value, is_vec),
-        "string" => {
-            // String is a special case because of quotes and to_string()
-            if is_vec {
-                // TODO there is a bug here, no idea how I should be attempting to convert / escape single quotes here...
-                let parsed: Vec<String> = serde_json::from_str(value).expect(
-            &format!("Failed to parse a literal value in a message file to the corresponding rust type: {value} to Vec<String>"));
-                let vec_str = format!("{parsed:?}.iter().map(|x| x.to_string()).collect()");
-                quote! { #vec_str }
-            } else {
-                // Halfass attempt to deal with ROS's string escaping / quote bullshit
-                let value = &value.replace("\'", "\"");
-                let parsed: String = serde_json::from_str(value).expect(&format!("Failed to parse a literal value in a message file to the corresponding rust type: {value} to String"));
-                quote! { #parsed }
-            }
-        }
-        _ => {
-            panic!("Found default for type which does not support default: {ros_type}");
         }
     }
 }