+++ CHANGELOG.md
@@ -23,10 +23,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Added public APIs for getting message data from search and for generating Rust code given message data in roslibrust_codegen
  - More useful logs available when running codegen
  - Refactor some of the public APIs and types in roslibrust_codegen (concept of `ParsedMessageFile` vs `MessageFile`)
+ - Added a method `get_md5sum` to `MessageFile`
+ - Additional code generation API and macro which excludes `ROS_PACKAGE_PATH`
 
 ### Fixed
  - Bug causing single quoted string constants in message files to not be parsed correctly
  - Bug causing float constants in message files to cause compiler errors because `f32 = 0;` is not allowed in rust
+ - Bug where packages were not properly de-duplicated during discovery.
 
 ## 0.6.0 - December 12, 2022